// src/runtime/IR.d.ts
export type IRDocument = ServerIRDocument | ClientIRDocument

export type BaseIRDocument = {
  // IR版本号
  ir_version: 1
  // IR类型标识
  ir_type: 'node_graph'
  // 节点图变量定义
  variables?: Variable[]
}

export type ServerIRDocument = SimplifyDeep<
  BaseIRDocument & {
    graph: ServerGraphInfo
    nodes?: ServerNode[]
  }
>

export type ClientIRDocument = SimplifyDeep<
  BaseIRDocument & {
    graph: ClientGraphInfo
    nodes?: ClientNode[]
  }
>

export type ServerGraphSubType = 'entity' | 'status' | 'class' | 'item'

export interface ServerGraphInfo {
  // 默认基于脚本文件名
  name?: string
  id?: number
  type: 'server'
  sub_type?: ServerGraphSubType
}

export interface ClientGraphInfo {
  name?: string
  id?: number
  type: 'client'
}
// ...
export type ServerNode = SimplifyDeep<
  Node & {
    type: string
  }
>

export type ClientNode = SimplifyDeep<
  Node & {
    type: string
  }
>

export type Argument =
  | ConnectionArgument
  | {
      [K in keyof ValueTypeMap]: {
        type: K
        value: ValueTypeMap[K]
      }
    }[keyof ValueTypeMap]
  | null

export interface ConnectionArgument {
  type: 'conn'
  value: ConnectionArgumentValue
}

export interface ConnectionArgumentValue {
  node_id: number
  index: number
  /**
   * Data value type carried by this connection.
   */
  type: ValueType
  /**
   * Extra type info for enum/enumeration connections.
   * Required when `type === 'enum'` or `type === 'enumeration'`.
   */
  enum?: string
  /**
   * Extra type info for dict connections.
   * Required when `type === 'dict'`.
   */
  dict?: {
    k: DictKeyType
    v: DictValueType
  }
  sub_index?: number
}

export type NextConnection = number | NextConnectionDetailed

export interface NextConnectionDetailed {
  // 目标节点ID
  node_id: number
  // 源节点的执行输出索引（可选）
  source_index?: number
  // 源节点的执行输出分支索引（可选）
  source_sub_index?: number
  // 目标节点的执行输入索引（可选）
  target_index?: number
  // 目标节点的执行输入分支索引（可选）
  target_sub_index?: number
}

export type ValueTypeMap = SimplifyDeep<LiteralValueTypeMap & SpecialValueTypeMap>

export type LiteralValueTypeMap = SimplifyDeep<
  LiteralValueTypeMapWithoutList & LiteralValueListTypeMap
>
// ...
export type SpecialValueTypeMap = SimplifyDeep<
  StructValueTypeMap &
    DictValueTypeMap &
    GenericValueTypeMap &
    EnumValueTypeMap &
    LocalVariableValueTypeMap &
    CustomVariableSnapshotValueTypeMap
>

export interface StructValueTypeMap {
  struct: any
}

export interface DictValueTypeMap {
  dict: any
}

export interface GenericValueTypeMap {
  generic: any
}

export interface EnumValueTypeMap {
  enum: string
  enumeration: string
}

export interface LocalVariableValueTypeMap {
  local_variable: any
}

export interface CustomVariableSnapshotValueTypeMap {
  custom_variable_snapshot: any
}

export type ValueType = SimplifyDeep<keyof ValueTypeMap>